+++ R/utils-is.R
@@ -2,7 +2,7 @@
 #'
 #' @noRd
 is_class <- function(x, classes = NULL, allow_null = FALSE) {
-  if (is.null(x) && allow_null) {
+  if (is.null(x) && isTRUE(allow_null)) {
     return(TRUE)
   }